refactor: add slices for activity and timer to separate logics
